Longest Substring solution with bit older class refactoring
