+++ src/main/kotlin/Main.kt
@@ -1,6 +1,7 @@
 package main.kotlin
 
-import TwoSum
+import main.kotlin.dsalgoleetcode.LongestSubstring
+import main.kotlin.dsalgoleetcode.TwoSum
 
 fun main() {
     println("What's your name?")
@@ -8,8 +9,13 @@ fun main() {
     println("Hello, $name!")
 
     // Calling Two Sum class
-    val obj = TwoSum()
-    val result = obj.twoSum(nums = intArrayOf(1,2,3,4,5), target = 9)
+    val twoSum = TwoSum()
+    val result = twoSum.twoSum(nums = intArrayOf(1, 2, 3, 4, 5), target = 9)
+
+    // Calling Longest Substring
+    val longestSubstring = LongestSubstring()
+    val lsResult = longestSubstring.lengthOfLongestSubstring(s = "abcabcbbb")
 
     println("Indices: ${result.joinToString(",")}")
+    println("Longest Substring : $lsResult")
 }
